Refactor: Lv3. Kiosk 클래스로 제어 담당 이전
