+++ src/com/sparta/kiosk/Main.java
@@ -4,6 +4,6 @@ public class Main {
 
   public static void main(String[] args) {
     Kiosk kiosk = new Kiosk();
-    kiosk.run();
+    kiosk.start();
   }
 }
